move exception handelr to a separte file
